no more magic action types
