+++ lib/components/AddItem/index.js
@@ -1,60 +1,18 @@
-import React, { Component } from 'react'
-import { TextInput, StyleSheet } from 'react-native'
-import ListItem from '../ListItem'
-import Colors from '../../constants/Colors'
+import AddItem from './AddItem'
+import {
+  ITEMS_CREATE_PENDING,
+  ITEMS_CREATE_SUCCESS,
+  ITEMS_CREATE_FAILURE,
+  ADD_ITEM_CHANGE,
+  ADD_ITEM_CLEAR
+} from '../../state/actions'
+import {
+
+} from '../../state/actions'
+import
 import { connect } from 'react-redux'
 
-class AddItem extends Component {
-  componentDidMount() {
-    const { addItemTextInput } = this.refs
-    const { addingNewItem } = this.props
-    if (addItemTextInput && addingNewItem) {
-      addItemTextInput.focus()
-    }
-  }
-  componentDidUpdate() {
-    const { addItemTextInput } = this.refs
-    const { addingNewItem } = this.props
-    if (addItemTextInput) {
-      if (addingNewItem) {
-        return addItemTextInput.focus()
-      }
-      addItemTextInput.blur()
-    }
-  }
-  render() {
-    const {text, handleChangeText, handleSubmitEditing, addingNewItem} = this.props
-    if (!addingNewItem) {
-      return null
-    }
-    return (
-      <ListItem
-        item={{}}
-        isEditing={true}
-      >
-        <TextInput
-          ref='addItemTextInput'
-          style={styles.textInput}
-          placeholder='What do you plan to do?'
-          value={text}
-          onChangeText={handleChangeText}
-          onSubmitEditing={handleSubmitEditing(text)}
-        />
-      </ListItem>
-    )
-  }
-}
-
-const styles = StyleSheet.create({
-  textInput: {
-    width: '100%',
-    height: '100%',
-    backgroundColor: Colors.white
-  }
-})
-
 const mapStateToProps = (state) => {
-  console.log('mapStateToProps')
   const {
     forms: {
       addItem: {
@@ -68,11 +26,12 @@ const mapStateToProps = (state) => {
     addingNewItem
   }
 }
+
 const mapDispatchToProps = (dispatch) => {
   return {
     handleChangeText: (text) => {
       dispatch({
-        type: 'ADD_ITEM_CHANGE',
+        type: ADD_ITEM_CHANGE,
         data: {
           text
         }
@@ -81,9 +40,9 @@ const mapDispatchToProps = (dispatch) => {
     handleSubmitEditing: (text) => () => {
       dispatch({
         types: [
-          'ITEMS_CREATE_PENDING',
-          'ITEMS_CREATE_SUCCESS',
-          'ITEMS_CREATE_FAILURE'
+          ITEMS_CREATE_PENDING,
+          ITEMS_CREATE_SUCCESS,
+          ITEMS_CREATE_FAILURE
         ],
         method: 'POST',
         data: {
@@ -91,11 +50,12 @@ const mapDispatchToProps = (dispatch) => {
         }
       })
       dispatch({
-        type: 'ADD_ITEM_CLEAR'
+        type: ADD_ITEM_CLEAR
       })
     }
   }
 }
+
 export default connect(
   mapStateToProps,
   mapDispatchToProps
