+++ lib/components/FloatingButton/index.js
@@ -1,4 +1,5 @@
 import FloatingButton from './FloatingButton'
+import { TOGGLE_ADDING_NEW_ITEM } from '../../state/actions'
 import { connect } from 'react-redux'
 
 const mapStateToProps = (state) => {
@@ -14,7 +15,7 @@ const mapDispatchToProps = (dispatch) => {
   return {
     handlePress: () => {
       dispatch({
-        type: 'TOGGLE_ADDING_NEW_ITEM',
+        type: TOGGLE_ADDING_NEW_ITEM,
         data: {}
       })
     }
