+++ lib/components/NavTab/index.js
@@ -1,4 +1,5 @@
 import NavTab from './NavTab'
+import { SWITCH_TAB } from '../../state/actions'
 import { connect } from 'react-redux'
 
 const mapStateToProps = (state) => {
@@ -14,7 +15,7 @@ const mapDispatchToProps = (dispatch) => {
   return {
     handleSelectTab: (tab) => () => {
       dispatch({
-        type: 'SWITCH_TAB',
+        type: SWITCH_TAB,
         data: { tab }
       })
     }
