+++ lib/components/TodoList/index.js
@@ -1,5 +1,13 @@
 import TodoList from './TodoList'
 import Tabs from '../../constants/Tabs'
+import {
+  ITEMS_UPDATE_PENDING,
+  ITEMS_UPDATE_SUCCESS,
+  ITEMS_UPDATE_FAILURE,
+  ITEMS_DELETE_PENDING,
+  ITEMS_DELETE_SUCCESS,
+  ITEMS_DELETE_FAILURE
+}
 import { connect } from 'react-redux'
 
 const mapStateToProps = (state) => {
@@ -37,9 +45,9 @@ const mapDispatchToProps = (dispatch) => {
       console.log(`toggleCompleted(${id}, ${completed})`)
       dispatch({
         types: [
-          'ITEMS_UPDATE_PENDING',
-          'ITEMS_UPDATE_SUCCESS',
-          'ITEMS_UPDATE_FAILURE'
+          ITEMS_UPDATE_PENDING,
+          ITEMS_UPDATE_SUCCESS,
+          ITEMS_UPDATE_FAILURE
         ],
         method: 'put',
         data: {
@@ -51,9 +59,9 @@ const mapDispatchToProps = (dispatch) => {
     deleteItem: (id) => () => {
       dispatch({
         types: [
-          'ITEMS_DELETE_PENDING',
-          'ITEMS_DELETE_SUCCESS',
-          'ITEMS_DELETE_FAILURE'
+          ITEMS_DELETE_PENDING,
+          ITEMS_DELETE_SUCCESS,
+          ITEMS_DELETE_FAILURE
         ],
         method: 'delete',
         data: { id }
